# schemas/openapi.v1.json
{
  "components": {
    "schemas": {
      "AltitudeConstraint": {
        "additionalProperties": false,
        "properties": {
          "foot": {
            "title": "Foot",
            "type": "string"
          },
          "from_point": {
            "title": "From Point",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "segment": {
            "title": "Segment",
            "type": "string"
          },
          "to_object": {
            "title": "To Object",
            "type": "string"
          },
          "type": {
            "const": "altitude",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "from_point",
          "to_object",
          "foot",
          "segment"
        ],
        "title": "AltitudeConstraint",
        "type": "object"
      },
      "AngleBisectorConstraint": {
        "additionalProperties": false,
        "properties": {
          "angle": {
            "title": "Angle",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "ray": {
            "title": "Ray",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "type": {
            "const": "angle_bisector",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "angle",
          "ray"
        ],
        "title": "AngleBisectorConstraint",
        "type": "object"
      },
      "AngleObject": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "points": {
            "maxItems": 3,
            "minItems": 3,
            "prefixItems": [
              {
                "type": "string"
              },
              {
                "type": "string"
              },
              {
                "type": "string"
              }
            ],
            "title": "Points",
            "type": "array"
          },
          "type": {
            "const": "angle",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "points"
        ],
        "title": "AngleObject",
        "type": "object"
      },
...
      "BelongsToConstraint": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "object": {
            "title": "Object",
            "type": "string"
          },
          "point": {
            "title": "Point",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "type": {
            "const": "belongs_to",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "point",
          "object"
        ],
        "title": "BelongsToConstraint",
        "type": "object"
      },
...
      "CircleObject": {
        "additionalProperties": false,
        "properties": {
          "center": {
            "title": "Center",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "radius_point": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Radius Point"
          },
          "type": {
            "const": "circle",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "center"
        ],
        "title": "CircleObject",
        "type": "object"
      },
      "CircumcircleConstraint": {
        "additionalProperties": false,
        "properties": {
          "circle": {
            "title": "Circle",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "triangle": {
            "title": "Triangle",
            "type": "string"
          },
          "type": {
            "const": "circumcircle",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "triangle",
          "circle"
        ],
        "title": "CircumcircleConstraint",
        "type": "object"
      },
      "CollinearConstraint": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "points": {
            "items": {
              "type": "string"
            },
            "title": "Points",
            "type": "array"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "type": {
            "const": "collinear",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "points"
        ],
        "title": "CollinearConstraint",
        "type": "object"
      },
      "ConstructionStep": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "constraints": {
            "items": {
              "type": "string"
            },
            "title": "Constraints",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "objects": {
            "items": {
              "type": "string"
            },
            "title": "Objects",
            "type": "array"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "id",
          "action",
          "objects"
        ],
        "title": "ConstructionStep",
        "type": "object"
      },
      "EqualLengthConstraint": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "objects": {
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "type": "string"
              },
              {
                "type": "string"
              }
            ],
            "title": "Objects",
            "type": "array"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "type": {
            "const": "equal_length",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "objects"
        ],
        "title": "EqualLengthConstraint",
        "type": "object"
      },
...
      "GirScene": {
        "$id": "urn:geometryos:gir:0.2.0",
        "additionalProperties": false,
        "properties": {
          "constraints": {
            "items": {
              "discriminator": {
                "mapping": {
                  "altitude": "#/components/schemas/AltitudeConstraint",
                  "angle_bisector": "#/components/schemas/AngleBisectorConstraint",
                  "belongs_to": "#/components/schemas/BelongsToConstraint",
                  "circumcircle": "#/components/schemas/CircumcircleConstraint",
                  "collinear": "#/components/schemas/CollinearConstraint",
                  "equal_length": "#/components/schemas/EqualLengthConstraint",
                  "incircle": "#/components/schemas/IncircleConstraint",
                  "intersection": "#/components/schemas/IntersectionConstraint",
                  "median": "#/components/schemas/MedianConstraint",
                  "midpoint": "#/components/schemas/MidpointConstraint",
                  "non_collinear": "#/components/schemas/NonCollinearConstraint",
                  "parallel": "#/components/schemas/ParallelConstraint",
                  "perpendicular": "#/components/schemas/PerpendicularConstraint"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/BelongsToConstraint"
                },
                {
                  "$ref": "#/components/schemas/CollinearConstraint"
                },
                {
                  "$ref": "#/components/schemas/NonCollinearConstraint"
                },
                {
                  "$ref": "#/components/schemas/ParallelConstraint"
                },
                {
                  "$ref": "#/components/schemas/PerpendicularConstraint"
                },
                {
                  "$ref": "#/components/schemas/EqualLengthConstraint"
                },
                {
                  "$ref": "#/components/schemas/MidpointConstraint"
                },
                {
                  "$ref": "#/components/schemas/IntersectionConstraint"
                },
                {
                  "$ref": "#/components/schemas/AltitudeConstraint"
                },
                {
                  "$ref": "#/components/schemas/MedianConstraint"
                },
                {
                  "$ref": "#/components/schemas/AngleBisectorConstraint"
                },
                {
                  "$ref": "#/components/schemas/CircumcircleConstraint"
                },
                {
                  "$ref": "#/components/schemas/IncircleConstraint"
                }
              ]
            },
            "title": "Constraints",
            "type": "array"
          },
          "construction_steps": {
            "items": {
              "$ref": "#/components/schemas/ConstructionStep"
            },
            "title": "Construction Steps",
            "type": "array"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "objects": {
            "items": {
              "discriminator": {
                "mapping": {
                  "angle": "#/components/schemas/AngleObject",
                  "circle": "#/components/schemas/CircleObject",
                  "label": "#/components/schemas/LabelObject",
                  "line": "#/components/schemas/LineObject",
                  "point": "#/components/schemas/PointObject",
                  "ray": "#/components/schemas/RayObject",
                  "segment": "#/components/schemas/SegmentObject",
                  "triangle": "#/components/schemas/TriangleObject"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/PointObject"
                },
                {
                  "$ref": "#/components/schemas/SegmentObject"
                },
                {
                  "$ref": "#/components/schemas/LineObject"
                },
                {
                  "$ref": "#/components/schemas/RayObject"
                },
                {
                  "$ref": "#/components/schemas/CircleObject"
                },
                {
                  "$ref": "#/components/schemas/TriangleObject"
                },
                {
                  "$ref": "#/components/schemas/AngleObject"
                },
                {
                  "$ref": "#/components/schemas/LabelObject"
                }
              ]
            },
            "title": "Objects",
            "type": "array"
          },
          "scene_type": {
            "const": "2d",
            "title": "Scene Type",
            "type": "string"
          },
          "schema_version": {
            "const": "0.2.0",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "scene_type",
          "objects",
          "constraints",
          "construction_steps"
        ],
        "title": "GirScene",
        "type": "object",
        "x-gir-schema-version": "0.2.0"
      },
      "IncircleConstraint": {
        "additionalProperties": false,
        "properties": {
          "circle": {
            "title": "Circle",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "triangle": {
            "title": "Triangle",
            "type": "string"
          },
          "type": {
            "const": "incircle",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "triangle",
          "circle"
        ],
        "title": "IncircleConstraint",
        "type": "object"
      },
      "IntersectionConstraint": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "objects": {
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "type": "string"
              },
              {
                "type": "string"
              }
            ],
            "title": "Objects",
            "type": "array"
          },
          "point": {
            "title": "Point",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "type": {
            "const": "intersection",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "point",
          "objects"
        ],
        "title": "IntersectionConstraint",
        "type": "object"
      },
      "LabelObject": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "target": {
            "title": "Target",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "type": {
            "const": "label",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "text",
          "target"
        ],
        "title": "LabelObject",
        "type": "object"
      },
      "LineObject": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "points": {
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "type": "string"
              },
              {
                "type": "string"
              }
            ],
            "title": "Points",
            "type": "array"
          },
          "type": {
            "const": "line",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "points"
        ],
        "title": "LineObject",
        "type": "object"
      },
      "MedianConstraint": {
        "additionalProperties": false,
        "properties": {
          "from_point": {
            "title": "From Point",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "midpoint": {
            "title": "Midpoint",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "segment": {
            "title": "Segment",
            "type": "string"
          },
          "to_object": {
            "title": "To Object",
            "type": "string"
          },
          "type": {
            "const": "median",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "from_point",
          "to_object",
          "midpoint",
          "segment"
        ],
        "title": "MedianConstraint",
        "type": "object"
      },
      "MidpointConstraint": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "object": {
            "title": "Object",
            "type": "string"
          },
          "point": {
            "title": "Point",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "type": {
            "const": "midpoint",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "point",
          "object"
        ],
        "title": "MidpointConstraint",
        "type": "object"
      },
      "NonCollinearConstraint": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "points": {
            "maxItems": 3,
            "minItems": 3,
            "prefixItems": [
              {
                "type": "string"
              },
              {
                "type": "string"
              },
              {
                "type": "string"
              }
            ],
            "title": "Points",
            "type": "array"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "type": {
            "const": "non_collinear",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "points"
        ],
        "title": "NonCollinearConstraint",
        "type": "object"
      },
      "ParallelConstraint": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "objects": {
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "type": "string"
              },
              {
                "type": "string"
              }
            ],
            "title": "Objects",
            "type": "array"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "type": {
            "const": "parallel",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "objects"
        ],
        "title": "ParallelConstraint",
        "type": "object"
      },
      "PerpendicularConstraint": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "objects": {
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "type": "string"
              },
              {
                "type": "string"
              }
            ],
            "title": "Objects",
            "type": "array"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "type": {
            "const": "perpendicular",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "objects"
        ],
        "title": "PerpendicularConstraint",
        "type": "object"
      },
...
      "ProblemDetail": {
        "additionalProperties": false,
        "examples": [
          {
            "code": "request_validation_failed",
            "detail": "The request payload does not satisfy the API contract.",
            "errors": [
              {
                "code": "literal_error",
                "location": [
                  "body",
                  "mode"
                ],
                "message": "Input should be 'strict'"
              }
            ],
            "instance": "/api/v1/generate",
            "request_id": "tutorboard-contract",
            "status": 422,
            "title": "Request validation failed",
            "type": "urn:geometryos:problem:request-validation"
          },
          {
            "code": "operation_timeout",
            "detail": "The generate operation exceeded its configured time limit.",
            "errors": [],
            "instance": "/api/v1/generate",
            "request_id": "tutorboard-contract",
            "status": 504,
            "title": "Operation timed out",
            "type": "urn:geometryos:problem:operation-timeout"
          },
          {
            "code": "internal_error",
            "detail": "An unexpected internal error occurred.",
            "errors": [],
            "instance": "/api/v1/generate",
            "request_id": "tutorboard-contract",
            "status": 500,
            "title": "Internal server error",
            "type": "urn:geometryos:problem:internal-error"
          }
        ],
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ProblemError"
            },
            "title": "Errors",
            "type": "array"
          },
          "instance": {
            "title": "Instance",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "request_id"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
...
      "RenderSvgV1Response": {
        "additionalProperties": false,
        "examples": [
          {
            "content": "<svg xmlns=\"http://www.w3.org/2000/svg\">...</svg>",
            "media_type": "image/svg+xml",
            "schema_version": "0.2.0"
          }
        ],
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "media_type": {
            "const": "image/svg+xml",
            "default": "image/svg+xml",
            "title": "Media Type",
            "type": "string"
          },
          "schema_version": {
            "const": "0.2.0",
            "default": "0.2.0",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "content"
        ],
        "title": "RenderSvgV1Response",
        "type": "object"
      },
      "RenderTikzV1Response": {
        "additionalProperties": false,
        "examples": [
          {
            "content": "\\begin{tikzpicture}...\\end{tikzpicture}",
            "media_type": "text/x-tex",
            "schema_version": "0.2.0"
          }
        ],
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "media_type": {
            "const": "text/x-tex",
            "default": "text/x-tex",
            "title": "Media Type",
            "type": "string"
          },
          "schema_version": {
            "const": "0.2.0",
            "default": "0.2.0",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "content"
        ],
        "title": "RenderTikzV1Response",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "GIR-first geometry compiler service.",
    "title": "GeometryOS API",
    "version": "1.0.0",
    "x-geometryos-api-major": "v1",
    "x-geometryos-consumer-contract": "tutorboard/v1",
    "x-geometryos-gir-schema-version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/render/svg": {
      "post": {
        "description": "Validate and normalize a GIR scene, then return deterministic SVG content.",
        "operationId": "geometryos_v1_render_svg",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GirScene"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenderSvgV1Response"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Request or GIR validation failed."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unexpected internal error."
          },
          "504": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Operation exceeded its configured time limit."
          }
        },
        "summary": "Render validated GIR as SVG",
        "tags": [
          "Rendering"
        ]
      }
    },
    "/api/v1/render/tikz": {
      "post": {
        "description": "Validate and normalize a GIR scene, then return deterministic TikZ content.",
        "operationId": "geometryos_v1_render_tikz",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GirScene"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenderTikzV1Response"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Request or GIR validation failed."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unexpected internal error."
          },
          "504": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Operation exceeded its configured time limit."
          }
        },
        "summary": "Render validated GIR as TikZ",
        "tags": [
          "Rendering"
        ]
      }
    },
...
  },
...
}